+++ coopr-provisioner/worker/plugins/automators/chef_solo_automator/resources/cookbooks/aufs/files/default/tests/minitest/module_test.rb
@@ -1,0 +1,9 @@
+require File.expand_path('../support/helpers', __FILE__)
+
+describe_recipe 'aufs::module' do
+  include Helpers::Aufs
+
+  it 'enables aufs module' do
+    assert_sh('modprobe -n -v aufs')
+  end
+end