+++ coopr-provisioner/worker/plugins/automators/chef_solo_automator/resources/cookbooks/device-mapper/recipes/default.rb
@@ -1,0 +1,3 @@
+include_recipe 'apt' if node['platform'] == 'ubuntu'
+include_recipe 'device-mapper::package'
+include_recipe 'device-mapper::module'