+++ coopr-provisioner/worker/plugins/automators/chef_solo_automator/resources/cookbooks/homebrew/recipes/default.rb
@@ -27,7 +27,7 @@ owner = homebrew_owner
 Chef::Log.debug("Homebrew owner is '#{homebrew_owner}'")
 
 remote_file homebrew_go do
-  source 'https://raw.github.com/Homebrew/homebrew/go/install'
+  source 'https://raw.githubusercontent.com/Homebrew/install/master/install'
   mode 00755
 end
 
@@ -36,11 +36,13 @@ execute homebrew_go do
   not_if { ::File.exist? '/usr/local/bin/brew' }
 end
 
-package 'git' do
-  not_if 'which git'
-end
+if node['homebrew']['auto-update']
+  package 'git' do
+    not_if 'which git'
+  end
 
-execute 'update homebrew from github' do
-  user owner
-  command '/usr/local/bin/brew update || true'
+  execute 'update homebrew from github' do
+    user owner
+    command '/usr/local/bin/brew update || true'
+  end
 end